Rename exec.hpp to query.hpp
