+++ include/pgasio/README.md
@@ -1,11 +1,11 @@
 # pgasio Headers
 
-To get started look at [`connection.hpp`](./connection.hpp#L9) to connect to the database and then at [`exec.hpp`](./exec.hpp#L9) to see how to send a `SELECT` to Postgres and traverse the resulting record set. [`csj.cpp`](./../../examples/csj.cpp) contains a full example.
+To get started look at [`connection.hpp`](./connection.hpp#L9) to connect to the database and then at [`query.hpp`](./query.hpp#L9) to see how to send a `SELECT` to Postgres and traverse the resulting record set. [`csj.cpp`](./../../examples/csj.cpp) contains a full example.
 
 For simple interactions with Postgres the general process is:
 
 1. Create a connection --  [`connection.hpp`](./connection.hpp#L9)
-2. Issue a command to Postgres -- [`exec.hpp`](./exec.hpp#L9), or [`network.hpp`](./network.hpp#L9) then possibly [`record_block.hpp`](./record_block.hpp#L9).
+2. Issue a command to Postgres -- [`query.hpp`](./query.hpp#L9), or [`network.hpp`](./network.hpp#L9) then possibly [`record_block.hpp`](./record_block.hpp#L9).
 3. Iterate through the messages Postgres returns until a `Z` packet is returned -- you can then return to step 2 if you need to.
 
 It's important that the resultant data is properly iterated though because the network traffic must be fully read.
@@ -41,16 +41,11 @@ Contains classes that can be used as either error returns or exceptions.
 * `end_of_packet` -- Thrown by the `decoder` when there aren't enough bytes left in the message.
 
 
-## [`exec.hpp`](./exec.hpp#L9)
-
-The `exec` function can be used to send a query to Postgres. This returns a `resultset` instance from which `recordset`s can be fetched. The `recordset` will deliver `record_block` instances (see [`record_block.hpp`](./record_block.hpp#L9)) from which rows can be decoded using the information in the `columne_meta` structures held by the `recordset`.
-
-
 ## [`memory.hpp`](./memory.hpp#L9)
 
 Contains an `array_view` implementation that can be used to share and manipulate contiguous blocks of memory without owning them. There are specialisations for byte arrays: `byte_view` is immutable and `raw_memory` is mutable. Wherever possible other interfaces are specified in terms of `array_view`s, see `decoder` for an example.
 
-`unaligned_slab` is an owner for a block of memory that unaligned sections can be fetched from. These sub-parts can be safely used as
+`unaligned_slab` is an owner for a block of memory that unaligned memory blocks can be allocated from.
 
 
 ## [`network.hpp`](./network.hpp#L9)
@@ -74,4 +69,9 @@ For an example of how these can be used together see the `handshake` implementat
 ## [`record_block.hpp`](./record_block.hpp#L9)
 
 Contains a block of record data fetched from a socket. It owns a large slab of memory (an `unaligned_slab`) together with a vector of field data locations so the column data can be interpreted correctly.
+
+
+## [`query.hpp`](./query.hpp#L9)
+
+The `query` function can be used to send a query to Postgres. This returns a `resultset` instance from which `recordset`s can be fetched. The `recordset` will deliver `record_block` instances (see [`record_block.hpp`](./record_block.hpp#L9)) from which rows can be decoded using the information in the `columne_meta` structures held by the `recordset`.
